# xchange-idexjaxrs/src/main/gen/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0",
    "title" : "idexjaxrs"
  },
  "host" : "api.idexjaxrs.market",
  "basePath" : "/",
...
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/returnOpenOrders" : {
      "post" : {
        "tags" : [ "trade" ],
        "summary" : "Returns the open orders for a given market and address",
        "operationId" : "OpenOrders",
        "parameters" : [ {
          "in" : "body",
          "name" : "OpenOrdersReq",
          "description" : "a JSON object the base market followed by an underscore, followed by the trade market",
          "required" : false,
          "schema" : {
            "type" : "object",
            "properties" : {
              "address" : {
                "type" : "string",
                "description" : "(address string) - Address to return open orders associated with"
              },
              "market" : {
                "type" : "string"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "schema" : {
              "$ref" : "#/definitions/returnOpenOrdersResponse"
            }
          }
        }
      }
    },
...
    "/returnOrderTrades" : {
      "post" : {
        "tags" : [ "trade" ],
        "summary" : "Returns all trades involving a given order hash, specified by the orderHash property of the JSON input.",
        "operationId" : "OrderTrades",
        "parameters" : [ {
          "in" : "body",
          "name" : "OrderTradesReq",
          "required" : false,
          "schema" : {
            "type" : "object",
            "properties" : {
              "orderHash" : {
                "type" : "string",
                "description" : "(256-bit hex string) - The order hash to query for associated trades"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Sample output: [ { date: '2017-10-11 21:41:15',amount: '0.3',type: 'buy',total: '1',price: '0.3',uuid: 'e8719a10-aecc-11e7-9535-3b8451fd4699',transactionHash: '0x28b945b586a5929c69337929533e04794d488c2d6e1122b7b915705d0dff8bb6' } ]",
            "schema" : {
              "$ref" : "#/definitions/returnOrderTradesResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "returnOpenOrdersResponse" : {
      "type" : "array",
      "items" : {
        "type" : "object",
        "properties" : {
          "orderNumber" : {
            "type" : "integer",
            "title" : "The Ordernumber Schema"
          },
          "orderHash" : {
            "type" : "string",
            "title" : "The Orderhash Schema",
            "default" : ""
          },
          "price" : {
            "type" : "string",
            "title" : "The Price Schema",
            "default" : ""
          },
          "amount" : {
            "type" : "string",
            "title" : "The Amount Schema",
            "default" : ""
          },
          "total" : {
            "type" : "string",
            "title" : "The Total Schema",
            "default" : ""
          },
          "type" : {
            "type" : "string",
            "title" : "The Type Schema",
            "default" : ""
          },
          "params" : {
            "type" : "object",
            "properties" : {
              "tokenBuy" : {
                "type" : "string",
                "title" : "The Tokenbuy Schema",
                "default" : ""
              },
              "buySymbol" : {
                "type" : "string",
                "title" : "The Buysymbol Schema",
                "default" : ""
              },
              "buyPrecision" : {
                "type" : "integer",
                "title" : "The Buyprecision Schema"
              },
              "amountBuy" : {
                "type" : "string",
                "title" : "The Amountbuy Schema",
                "default" : ""
              },
              "tokenSell" : {
                "type" : "string",
                "title" : "The Tokensell Schema",
                "default" : ""
              },
              "sellSymbol" : {
                "type" : "string",
                "title" : "The Sellsymbol Schema",
                "default" : ""
              },
              "sellPrecision" : {
                "type" : "integer",
                "title" : "The Sellprecision Schema"
              },
              "amountSell" : {
                "type" : "string",
                "title" : "The Amountsell Schema",
                "default" : ""
              },
              "expires" : {
                "type" : "integer",
                "title" : "The Expires Schema"
              },
              "nonce" : {
                "type" : "integer",
                "title" : "The Nonce Schema"
              },
              "user" : {
                "type" : "string",
                "title" : "The User Schema",
                "default" : ""
              }
            }
          }
        }
      }
    },
...
    "returnOrderTradesResponse" : {
      "type" : "array",
      "items" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "title" : "The Date Schema",
            "default" : ""
          },
          "amount" : {
            "type" : "string",
            "title" : "The Amount Schema",
            "default" : ""
          },
          "type" : {
            "type" : "string",
            "title" : "The Type Schema",
            "default" : ""
          },
          "total" : {
            "type" : "string",
            "title" : "The Total Schema",
            "default" : ""
          },
          "price" : {
            "type" : "string",
            "title" : "The Price Schema",
            "default" : ""
          },
          "uuid" : {
            "type" : "string",
            "title" : "The Uuid Schema",
            "default" : ""
          },
          "transactionHash" : {
            "type" : "string",
            "title" : "The Transactionhash Schema",
            "default" : ""
          }
        }
      }
    },
...
  }
}